+++ Sources/Sentry/Profiling/SentryProfilerSerialization.m
@@ -7,9 +7,7 @@
 #    import "SentryDebugImageProvider+HybridSDKs.h"
 #    import "SentryDependencyContainer.h"
 #    import "SentryDevice.h"
-#    import "SentryEnvelope.h"
 #    import "SentryEnvelopeItemHeader.h"
-#    import "SentryEnvelopeItemType.h"
 #    import "SentryEvent+Private.h"
 #    import "SentryFormatter.h"
 #    import "SentryInternalDefines.h"
@@ -346,13 +344,16 @@ SentryEnvelope *_Nullable sentry_continuousProfileChunkEnvelope(
 #    endif // defined(SENTRY_TEST) || defined(SENTRY_TEST_CI)
 
     SentryEnvelopeItemHeader *header =
-        [[SentryEnvelopeItemHeader alloc] initWithType:SentryEnvelopeItemTypeProfileChunk
+        [[SentryEnvelopeItemHeader alloc] initWithType:SentryEnvelopeItemTypes.profileChunk
                                                 length:JSONData.length];
     header.platform = @"cocoa";
     SentryEnvelopeItem *envelopeItem = [[SentryEnvelopeItem alloc] initWithHeader:header
                                                                              data:JSONData];
 
+#    pragma clang diagnostic push
+#    pragma clang diagnostic ignored "-Wdeprecated-declarations"
     return [[SentryEnvelope alloc] initWithId:chunkID singleItem:envelopeItem];
+#    pragma clang diagnostic pop
 }
 
 SentryEnvelopeItem *_Nullable sentry_traceProfileEnvelopeItem(SentryHub *hub,
@@ -398,7 +399,7 @@ SentryEnvelopeItem *_Nullable sentry_traceProfileEnvelopeItem(SentryHub *hub,
 #    endif // defined(SENTRY_TEST) || defined(SENTRY_TEST_CI)
 
     SentryEnvelopeItemHeader *header =
-        [[SentryEnvelopeItemHeader alloc] initWithType:SentryEnvelopeItemTypeProfile
+        [[SentryEnvelopeItemHeader alloc] initWithType:SentryEnvelopeItemTypes.profile
                                                 length:JSONData.length];
     return [[SentryEnvelopeItem alloc] initWithHeader:header data:JSONData];
 }