+++ Sources/Sentry/include/SentrySerialization.h
@@ -17,8 +17,6 @@ NS_ASSUME_NONNULL_BEGIN
 
 + (NSData *_Nullable)dataWithEnvelope:(SentryEnvelope *)envelope;
 
-+ (NSData *_Nullable)dataWithReplayRecording:(SentryReplayRecording *)replayRecording;
-
 + (SentryEnvelope *_Nullable)envelopeWithData:(NSData *)data;
 
 + (SentryAppState *_Nullable)appStateWithData:(NSData *)sessionData;