+++ Sources/Swift/Helper/SentryMobileProvisionParser.swift
@@ -52,14 +52,8 @@ public class SentryMobileProvisionParser: NSObject {
 
         var format = PropertyListSerialization.PropertyListFormat.xml
         
-#if swift(>=5.9) && os(visionOS)
-        let options = 0
-#else
-        let options: PropertyListSerialization.ReadOptions = []
-#endif
-        
         guard let obj = try? PropertyListSerialization.propertyList(from: plistData,
-                                                                    options: options,
+                                                                    options: [],
                                                                     format: &format),
               let dict = obj as? [String: Any] else {
             return