+++ Tests/SentryTests/Integrations/Breadcrumbs/SentrySystemEventBreadcrumbsTest.swift
@@ -14,7 +14,7 @@ class SentrySystemEventBreadcrumbsTest: XCTestCase {
         var currentDateProvider = TestCurrentDateProvider()
         let notificationCenterWrapper = TestNSNotificationCenterWrapper()
 
-        init() {
+        init() throws {
             options = Options()
             options.dsn = TestConstants.dsnAsString(username: "SentrySystemEventBreadcrumbsTest")
             options.releaseName = "SentrySessionTrackerIntegrationTests"
@@ -22,7 +22,11 @@ class SentrySystemEventBreadcrumbsTest: XCTestCase {
             options.environment = "debug"
             SentryDependencyContainer.sharedInstance().dateProvider = currentDateProvider
 
-            fileManager = try! TestFileManager(options: options)
+            fileManager = try TestFileManager(
+                options: options,
+                dateProvider: currentDateProvider,
+                dispatchQueueWrapper: TestSentryDispatchQueueWrapper()
+            )
         }
 
         func getSut(currentDevice: UIDevice? = UIDevice.current) -> SentrySystemEventBreadcrumbs {
@@ -61,9 +65,9 @@ class SentrySystemEventBreadcrumbsTest: XCTestCase {
     private var fixture: Fixture!
     private var sut: SentrySystemEventBreadcrumbs!
 
-    override func setUp() {
-        super.setUp()
-        fixture = Fixture()
+    override func setUpWithError() throws {
+        try super.setUpWithError()
+        fixture = try Fixture()
     }
 
     override func tearDown() {
@@ -324,10 +328,37 @@ class SentrySystemEventBreadcrumbsTest: XCTestCase {
         }
     }
 
+    func testSignificantTimeChangeNotificationBreadcrumb() throws {
+        sut = fixture.getSut(currentDevice: nil)
+
+        fixture.notificationCenterWrapper.post(Notification(name: UIApplication.significantTimeChangeNotification, object: nil))
+
+        XCTAssertEqual(1, fixture.delegate.addCrumbInvocations.count)
+
+        let crumb = try XCTUnwrap(fixture.delegate.addCrumbInvocations.first)
+
+        XCTAssertEqual("device.event", crumb.category)
+        XCTAssertEqual("system", crumb.type)
+        XCTAssertEqual(SentryLevel.info, crumb.level)
+
+        let data = try XCTUnwrap(crumb.data, "no breadcrumb.data")
+        XCTAssertEqual("SIGNIFICANT_TIME_CHANGE", data["action"] as? String)
+    }
+
+    func testSignificantTimeChangeNotificationBreadcrumb_UnsubscribeOnStop() {
+        sut = fixture.getSut(currentDevice: nil)
+
+        sut.stop()
+
+        let didCallRemoveObserver = fixture.notificationCenterWrapper.removeObserverWithNameAndObjectInvocations.invocations.filter { $0.name == UIApplication.significantTimeChangeNotification }.count == 1
+
+        XCTAssertTrue(didCallRemoveObserver, "Stop didn't call remove observer for UIApplicationSignificantTimeChangeNotification")
+    }
+
     func testStopCallsSpecificRemoveObserverMethods() {
         sut = fixture.getSut(currentDevice: nil)
         sut.stop()
-        XCTAssertEqual(fixture.notificationCenterWrapper.removeObserverWithNameAndObjectInvocations.count, 6)
+        XCTAssertEqual(fixture.notificationCenterWrapper.removeObserverWithNameAndObjectInvocations.count, 7)
     }
     
     private func postBatteryLevelNotification(uiDevice: UIDevice?) {